+++ ECom-app/PMS/src/main/java/com/itsthatjun/ecommerce/mbg/model/Product.java
@@ -20,17 +20,17 @@ public class Product {
 
     private String sn;
 
-    private BigDecimal newStatus;
+    private Integer newStatus;
 
-    private BigDecimal recommendStatus;
+    private Integer recommendStatus;
 
-    private BigDecimal verifyStatus;
+    private Integer verifyStatus;
 
     private String subTitle;
 
     private String coverPicture;
 
-    private BigDecimal pictureAlbum;
+    private Integer pictureAlbum;
 
     private String description;
 
@@ -128,27 +128,27 @@ public class Product {
         this.sn = sn;
     }
 
-    public BigDecimal getNewStatus() {
+    public Integer getNewStatus() {
         return newStatus;
     }
 
-    public void setNewStatus(BigDecimal newStatus) {
+    public void setNewStatus(Integer newStatus) {
         this.newStatus = newStatus;
     }
 
-    public BigDecimal getRecommendStatus() {
+    public Integer getRecommendStatus() {
         return recommendStatus;
     }
 
-    public void setRecommendStatus(BigDecimal recommendStatus) {
+    public void setRecommendStatus(Integer recommendStatus) {
         this.recommendStatus = recommendStatus;
     }
 
-    public BigDecimal getVerifyStatus() {
+    public Integer getVerifyStatus() {
         return verifyStatus;
     }
 
-    public void setVerifyStatus(BigDecimal verifyStatus) {
+    public void setVerifyStatus(Integer verifyStatus) {
         this.verifyStatus = verifyStatus;
     }
 
@@ -168,11 +168,11 @@ public class Product {
         this.coverPicture = coverPicture;
     }
 
-    public BigDecimal getPictureAlbum() {
+    public Integer getPictureAlbum() {
         return pictureAlbum;
     }
 
-    public void setPictureAlbum(BigDecimal pictureAlbum) {
+    public void setPictureAlbum(Integer pictureAlbum) {
         this.pictureAlbum = pictureAlbum;
     }
